IS1146 [filebrowser] Fixed XxdTest:test_compare_to_xxd failed (#1146)
- xxd_test.py : In this case, chr(not unichr) is enough for python2 and python3
- xxd.py: hex.ljust expect the first argument as integer not float

+++ apps/filebrowser/src/filebrowser/lib/xxd.py
@@ -94,7 +94,7 @@ def main(input, output):
         return "%02x" % ord
       hex = " ".join([ "".join(map(ashex, sentence)) for sentence in ordinals])
       # 2 characters per byte, 1 extra for spacing, and 1 extra at the end.
-      hex = hex.ljust(bytes_per_line*2 + (math.floor(bytes_per_line / bytes_per_sentence)) - 1)
+      hex = hex.ljust(bytes_per_line*2 + int(math.floor(bytes_per_line / bytes_per_sentence)) - 1)
       output.write("%07x: %s  %s\n" % (off, hex, printable))
 
     offset += len(data)
